feat: Introduce ActionButtons component to encapsulate sharing and download functionality; refactor page to utilize new component for improved code organization
