+++ src/app/result/create-image/components/index.ts
@@ -1,2 +1,3 @@
 export { default as Carousel } from "./Carousel";
 export { default as CarouselItem } from "./CarouselItem";
+export { default as ActionButtons } from "./ActionButtons";
